Marble Launcher - Inspired By SpinLaunch
Inspired by SpinLaunch, a company aiming to simplify space travel and make it more affordable. They aim to do this by removing traditional propulsion methods and instead using a propeller-like launch system with the payload on one side and a counter-mass on the other. While this system requires incredibly precise timing and a vacuum to remove air resistance, this basic model uses the walls themselves and a primitive "fork" to keep the ball in until it should be launched.