Closet Organization System
After looking in my closet one day, I realized that the current system was not working the way I would have liked. I took it upon myself to design a new closet organization system which would allow me to hang and organize my clothes to my liking, including giving room for my shoes. This model was the final design for the closet.